♻️ #101 refactor: 노트링크입력모달 분리
