+++ components/TiptapEditor.tsx
@@ -9,10 +9,13 @@ const TiptapEditor = ({ ...props }: PropsWithChildren) => {
   if (!editor) return;
 
   return (
-    <>
-      <EditorContent editor={editor} {...props} />
-      <div className='h-20'></div>
-    </>
+    <div className='grow lg:relative'>
+      <div className='overflow-visible h-full'>
+        <EditorContent editor={editor} {...props} />
+        <div className='h-20'></div>
+      </div>
+    </div>
   );
 };
+
 export default TiptapEditor;
